+++ extension-requests/index.html
@@ -103,7 +103,6 @@
     <h2 id="error"></h2>
     <script src="/constants.js"></script>
     <script src="/utils.js"></script>
-    <script src="local-utils.js"></script>
     <script src="script.js"></script>
   </body>
 </html>
