+++ online-members/online-members.html
@@ -10,8 +10,8 @@
     <link rel="stylesheet" href="/style.css" async />
     <link rel="stylesheet" href="/online-members/online-members.css" />
     <script src="/constants.js"></script>
+    <script src="/utils.js" defer></script>
     <script src="/online-members/sse-events.js" defer></script>
-    <script src="/online-members/utils.js" defer></script>
     <script src="/online-members/script.js" defer></script>
   </head>
