+++ taskEvents/index.html
@@ -16,6 +16,7 @@
     />
     <link rel="stylesheet" href="/taskEvents/style.css" />
     <script src="/constants.js"></script>
+    <script src="/utils.js"></script>
     <script src="/taskEvents/script.js" type="module" defer></script>
   </head>
