+++ users/details/index.html
@@ -10,7 +10,6 @@
     <link rel="stylesheet" href="/users/details/style.css" />
     <script src="/constants.js"></script>
     <script src="/utils.js"></script>
-    <script src="/users/details/utils.js" defer></script>
     <script src="/users/details/script.js" defer></script>
   </head>
