+++ users/discord/index.html
@@ -14,6 +14,7 @@
     </header>
     <div id="root"></div>
   </body>
+  <script src="/utils.js"></script>
   <script src="utils/react.js"></script>
   <script src="/constants.js"></script>
   <script type="module" src="index.js"></script>
