+++ users/index.html
@@ -10,7 +10,6 @@
     <link rel="stylesheet" href="/users/style.css" />
     <script src="/constants.js"></script>
     <script src="/utils.js"></script>
-    <script src="/users/utils.js" defer></script>
     <script type="module" src="/users/script.js" defer></script>
   </head>
