--- users/utils.js
@@ -1,35 +1,0 @@
-async function makeApiCall(
-  url,
-  method = 'get',
-  body = null,
-  credentials = 'include',
-  headers = { 'content-type': 'application/json' },
-  options = null,
-) {
-  try {
-    const response = await fetch(url, {
-      method,
-      body,
-      headers,
-      credentials,
-      ...options,
-    });
-    return response;
-  } catch (err) {
-    console.error(err);
-    throw err;
-  }
-}
-
-function debounce(func, delay) {
-  let timerId;
-  return (...args) => {
-    if (timerId) {
-      clearTimeout(timerId);
-    }
-
-    timerId = setTimeout(() => {
-      func(...args);
-    }, delay);
-  };
-}
